+++ Podfile
@@ -1,25 +1,25 @@
 # Uncomment the next line to define a global platform for your project
 platform :ios, '11.0'
 
-target 'SpokeStack' do
+target 'Spokestack' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
 
-  # Pods for SpokeStack
+  # Pods for Spokestack
   pod 'TensorFlowLiteSwift', '~> 1.14.0'
   pod 'filter_audio', '~> 0.4.3', :modular_headers => true
 
-  target 'SpokeStackTests' do
+  target 'SpokestackTests' do
     inherit! :search_paths
     # Pods for testing
   end
 
 end
 
-target 'SpokeStackFrameworkExample' do
+target 'SpokestackFrameworkExample' do
   # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
   use_frameworks!
 
-  # Pods for SpokeStackFrameworkExample
+  # Pods for SpokestackFrameworkExample
 
 end
