+++ SpokeStack.podspec
@@ -1,20 +1,20 @@
 Pod::Spec.new do |s|
-  s.name = 'SpokeStack'
-  s.version = '3.0.3'
+  s.name = 'Spokestack'
+  s.version = '4.0.0'
   s.license = 'Apache'
   s.summary = 'Spokestack provides an extensible speech recognition pipeline for the iOS platform.'
-  s.homepage = 'https://www.pylon.com'
-  s.authors = { 'Spokestack' => 'support@pylon.com' }
+  s.homepage = 'https://www.spokestack.io'
+  s.authors = { 'Spokestack' => 'support@spokestack.io' }
   s.source = { :git => 'https://github.com/pylon/spokestack-ios.git', :tag => s.version.to_s }
   s.license = {:type => 'Apache', :file => 'LICENSE'}
   s.ios.deployment_target = '11.0'
   s.swift_version = '4.2'
   s.ios.framework = 'AVFoundation', 'CoreML'
-  s.exclude_files = 'SpokeStackFrameworkExample/*.*', 'SpokeStackTests/*.*', 'SpokeStack/Info.plist'
-  s.source_files = 'SpokeStack/**/*.{swift,h,m,c}'
-  s.pod_target_xcconfig = {'SWIFT_INCLUDE_PATHS' => '$(SRCROOT)/SpokeStack/VAD/Wit', 'HEADER_SEARCH_PATHS' => '$(SRCROOT)/SpokeStack/VAD/Wit'}
-  s.preserve_paths = 'SpokeStack/**/*.modulemap'
-  s.public_header_files = 'SpokeStack/SpokeStack.h'
+  s.exclude_files = 'SpokestackFrameworkExample/*.*', 'SpokestackTests/*.*', 'Spokestack/Info.plist'
+  s.source_files = 'Spokestack/**/*.{swift,h,m,c}'
+  s.pod_target_xcconfig = {'SWIFT_INCLUDE_PATHS' => '$(SRCROOT)/Spokestack/VAD/Wit', 'HEADER_SEARCH_PATHS' => '$(SRCROOT)/Spokestack/VAD/Wit'}
+  s.preserve_paths = 'Spokestack/**/*.modulemap'
+  s.public_header_files = 'Spokestack/Spokestack.h'
   s.dependency 'TensorFlowLiteSwift', '~> 1.14.0'
   s.dependency 'filter_audio', '~> 0.4.3'
   s.static_framework = true
