+++ SpokeStack.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,12 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:SpokeStack.xcodeproj">
+      location = "group:/Users/noel/w/Spokestack-ios/Spokestack.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
+   </FileRef>
+   <FileRef
+      location = "group:Spokestack.xcodeproj">
    </FileRef>
 </Workspace>
