+++ SpokeStack/FFT.swift
@@ -1,6 +1,6 @@
 //
 //  FFT.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Cory D. Wiles on 12/6/18.
 //  Copyright © 2018 Pylon AI, Inc. All rights reserved.
