+++ SpokeStack/PipelineDelegate.swift
@@ -1,6 +1,6 @@
 //
 //  SpeechPipelineService.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Noel Weichbrodt on 3/21/19.
 //  Copyright © 2019 Pylon AI, Inc. All rights reserved.
