+++ SpokeStack/RingBuffer.swift
@@ -1,6 +1,6 @@
 //
 //  RingBuffer.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Cory D. Wiles on 12/5/18.
 //  Copyright © 2018 Pylon AI, Inc. All rights reserved.
