+++ SpokeStack/SignalProcessing.swift
@@ -1,6 +1,6 @@
 //
 //  SignalProcessingAlgorithms.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Noel Weichbrodt on 8/9/19.
 //  Copyright © 2019 Pylon AI, Inc. All rights reserved.
