+++ SpokeStack/SpeechContext.swift
@@ -1,6 +1,6 @@
 //
 //  SpeechContext.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Cory D. Wiles on 10/1/18.
 //  Copyright © 2018 Pylon AI, Inc. All rights reserved.
