+++ SpokeStack/SpokeStack.h
@@ -1,6 +1,6 @@
 //
-//  SpokeStack.h
-//  SpokeStack
+//  Spokestack.h
+//  Spokestack
 //
 //  Created by Cory D. Wiles on 10/8/18.
 //  Copyright © 2018 Pylon AI, Inc. All rights reserved.
@@ -8,12 +8,12 @@
 
 #import <UIKit/UIKit.h>
 
-//! Project version number for SpokeStack.
-FOUNDATION_EXPORT double SpokeStackVersionNumber;
+//! Project version number for Spokestack.
+FOUNDATION_EXPORT double SpokestackVersionNumber;
 
-//! Project version string for SpokeStack.
-FOUNDATION_EXPORT const unsigned char SpokeStackVersionString[];
+//! Project version string for Spokestack.
+FOUNDATION_EXPORT const unsigned char SpokestackVersionString[];
 
-// In this header, you should import all the public headers of your framework using statements like #import <SpokeStack/PublicHeader.h>
+// In this header, you should import all the public headers of your framework using statements like #import <Spokestack/PublicHeader.h>
