+++ SpokeStack/SpokeStackConfig.xcconfig
@@ -1,5 +1,5 @@
 //
-//  SpokeStackConfig.xcconfig
+//  SpokestackConfig.xcconfig
 //  IndependentFramework
 //
 //  Created by Cory D. Wiles on 10/5/18.
