+++ SpokeStack/TFLiteWakewordRecognizer.swift
@@ -1,6 +1,6 @@
 //
 //  TFLiteWakewordRecognizer.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Noel Weichbrodt on 8/5/19.
 //  Copyright © 2019 Pylon AI, Inc. All rights reserved.
