+++ SpokeStack/Typealiases.swift
@@ -1,6 +1,6 @@
 //
 //  Typealiases.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Cory D. Wiles on 9/28/18.
 //  Copyright © 2018 Pylon AI, Inc. All rights reserved.
