+++ SpokeStack/VADDelegate.swift
@@ -1,6 +1,6 @@
 //
 //  VADDelegate.swift
-//  SpokeStack
+//  Spokestack
 //
 //  Created by Noel Weichbrodt on 7/9/19.
 //  Copyright © 2019 Pylon AI, Inc. All rights reserved.
