+++ Spokestack.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
