+++ Spokestack.xcodeproj/xcshareddata/xcschemes/Spokestack.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "1010"
+   LastUpgradeVersion = "1000"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,10 +14,10 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "011C191C216BCA3200DB36FD"
-               BuildableName = "SpokeStackFrameworkExample.app"
-               BlueprintName = "SpokeStackFrameworkExample"
-               ReferencedContainer = "container:SpokeStack.xcodeproj">
+               BlueprintIdentifier = "01D31F3A216BAF260055FD45"
+               BuildableName = "Spokestack.framework"
+               BlueprintName = "Spokestack"
+               ReferencedContainer = "container:Spokestack.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -27,44 +27,47 @@
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES">
-      <Testables>
-      </Testables>
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "011C191C216BCA3200DB36FD"
-            BuildableName = "SpokeStackFrameworkExample.app"
-            BlueprintName = "SpokeStackFrameworkExample"
-            ReferencedContainer = "container:SpokeStack.xcodeproj">
+            BlueprintIdentifier = "01D31F3A216BAF260055FD45"
+            BuildableName = "Spokestack.framework"
+            BlueprintName = "Spokestack"
+            ReferencedContainer = "container:Spokestack.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
-      <AdditionalOptions>
-      </AdditionalOptions>
+      <Testables>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "01D31F43216BAF260055FD45"
+               BuildableName = "SpokestackTests.xctest"
+               BlueprintName = "SpokestackTests"
+               ReferencedContainer = "container:Spokestack.xcodeproj">
+            </BuildableReference>
+         </TestableReference>
+      </Testables>
    </TestAction>
    <LaunchAction
       buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
-      enableASanStackUseAfterReturn = "YES"
-      disableMainThreadChecker = "YES"
       launchStyle = "0"
       useCustomWorkingDirectory = "NO"
       ignoresPersistentStateOnLaunch = "NO"
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "011C191C216BCA3200DB36FD"
-            BuildableName = "SpokeStackFrameworkExample.app"
-            BlueprintName = "SpokeStackFrameworkExample"
-            ReferencedContainer = "container:SpokeStack.xcodeproj">
+            BlueprintIdentifier = "01D31F3A216BAF260055FD45"
+            BuildableName = "Spokestack.framework"
+            BlueprintName = "Spokestack"
+            ReferencedContainer = "container:Spokestack.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
-      <AdditionalOptions>
-      </AdditionalOptions>
+      </MacroExpansion>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -72,16 +75,15 @@
       savedToolIdentifier = ""
       useCustomWorkingDirectory = "NO"
       debugDocumentVersioning = "YES">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "011C191C216BCA3200DB36FD"
-            BuildableName = "SpokeStackFrameworkExample.app"
-            BlueprintName = "SpokeStackFrameworkExample"
-            ReferencedContainer = "container:SpokeStack.xcodeproj">
+            BlueprintIdentifier = "01D31F3A216BAF260055FD45"
+            BuildableName = "Spokestack.framework"
+            BlueprintName = "Spokestack"
+            ReferencedContainer = "container:Spokestack.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
    </ProfileAction>
    <AnalyzeAction
       buildConfiguration = "Debug">
