+++ SpokestackFrameworkExample/Assets.xcassets/AppIcon.appiconset/Contents.json
