+++ SpokestackFrameworkExample/Assets.xcassets/Contents.json
