+++ SpokestackFrameworkExample/Base.lproj/LaunchScreen.storyboard
