+++ SpokestackFrameworkExample/CoreMLViewController.swift
@@ -1,16 +1,16 @@
 //
-//  ViewController.swift
-//  SpokeStackFrameworkExample
+//  CoreMLViewController.swift
+//  SpokestackFrameworkExample
 //
-//  Created by Cory D. Wiles on 10/8/18.
-//  Copyright © 2018 Pylon AI, Inc. All rights reserved.
+//  Created by Noel Weichbrodt on 6/13/19.
+//  Copyright © 2019 Pylon AI, Inc. All rights reserved.
 //
 
 import UIKit
-import SpokeStack
+import Spokestack
 import AVFoundation
 
-class AppleASRViewController: UIViewController {
+class CoreMLViewController: UIViewController {
     
     lazy var startRecordingButton: UIButton = {
         
@@ -19,10 +19,9 @@ class AppleASRViewController: UIViewController {
         button.translatesAutoresizingMaskIntoConstraints = false
         button.setTitle("Start", for: .normal)
         button.addTarget(self,
-                         action: #selector(AppleASRViewController.startRecordingAction(_:)),
+                         action: #selector(CoreMLViewController.startRecordingAction(_:)),
                          for: .touchUpInside)
-        
-        button.setTitleColor(.blue, for: .normal)
+        button.setTitleColor(.purple, for: .normal)
         
         return button
     }()
@@ -34,36 +33,26 @@ class AppleASRViewController: UIViewController {
         button.translatesAutoresizingMaskIntoConstraints = false
         button.setTitle("Stop", for: .normal)
         button.addTarget(self,
-                         action: #selector(AppleASRViewController.stopRecordingAction(_:)),
+                         action: #selector(CoreMLViewController.stopRecordingAction(_:)),
                          for: .touchUpInside)
         
-        button.setTitleColor(.blue, for: .normal)
+        button.setTitleColor(.purple, for: .normal)
         
         
         return button
     }()
     
-    lazy private var pipeline: SpeechPipeline = {
-        
-        let appleConfiguration: SpeechConfiguration = SpeechConfiguration()
-        
-        return try! SpeechPipeline(SpeechProcessors.appleSpeech.processor,
-                                   speechConfiguration: appleConfiguration,
-                                   speechDelegate: self,
-                                   wakewordService: SpeechProcessors.appleWakeword.processor,
-                                   wakewordDelegate: self,
-                                   pipelineDelegate: self)
-    }()
+    private var pipeline: SpeechPipeline?
     
     override func loadView() {
         
         super.loadView()
         self.view.backgroundColor = .white
-        self.title = "Apple ASR"
+        self.title = "CoreML"
         
         let doneBarButtonItem: UIBarButtonItem = UIBarButtonItem(barButtonSystemItem: .done,
                                                                  target: self,
-                                                                 action: #selector(AppleASRViewController.dismissViewController(_:)))
+                                                                 action: #selector(CoreMLViewController.dismissViewController(_:)))
         self.navigationItem.rightBarButtonItem = doneBarButtonItem
     }
     
@@ -80,25 +69,56 @@ class AppleASRViewController: UIViewController {
         self.stopRecordingButton.topAnchor.constraint(equalTo: self.startRecordingButton.bottomAnchor, constant: 50.0).isActive = true
         self.stopRecordingButton.leftAnchor.constraint(equalTo: self.startRecordingButton.leftAnchor).isActive = true
         self.stopRecordingButton.rightAnchor.constraint(equalTo: self.startRecordingButton.rightAnchor).isActive = true
+        
+        do {
+            self.pipeline = try self.initPipeline()
+        } catch let error {
+            print("couldn't initialize pipeline becuase \(error)")
+        }
+    }
+    
+    func initPipeline() throws -> SpeechPipeline {
+        let c = SpeechConfiguration()
+        guard let filterPath = Bundle(for: type(of: self)).path(forResource: c.filterModelName, ofType: "mlmodelc") else {
+            throw WakewordModelError.filter("could not find \(c.filterModelName).mlmodelc in bundle \(self.debugDescription)")
+        }
+        c.filterModelPath = filterPath
+        guard let detectPath = Bundle(for: type(of: self)).path(forResource: c.detectModelName, ofType: "mlmodelc") else {
+            throw WakewordModelError.detect("could not find \(c.detectModelName).mlmodelc in bundle \(self.debugDescription)")
+        }
+        c.detectModelPath = detectPath
+        c.tracing = Trace.Level.PERF
+        return try! SpeechPipeline(SpeechProcessors.appleSpeech.processor,
+                                   speechConfiguration: c,
+                                   speechDelegate: self,
+                                   wakewordService: SpeechProcessors.coremlWakeword.processor,
+                                   wakewordDelegate: self,
+                                   pipelineDelegate: self)
     }
     
     @objc func startRecordingAction(_ sender: Any) {
         print("pipeline started")
-        self.pipeline.start()
+        self.pipeline?.start()
     }
     
     @objc func stopRecordingAction(_ sender: Any) {
         print("pipeline finished")
-        self.pipeline.stop()
+        self.pipeline?.stop()
     }
     
     @objc func dismissViewController(_ sender: Any?) -> Void {
         self.dismiss(animated: true, completion: nil)
     }
+    
+    func toggleStartStop() {
+        DispatchQueue.main.async {
+            self.stopRecordingButton.isEnabled.toggle()
+            self.startRecordingButton.isEnabled.toggle()
+        }
+    }
 }
 
-extension AppleASRViewController: SpeechEventListener, PipelineDelegate {
-    
+extension CoreMLViewController: SpeechEventListener, PipelineDelegate {
     func setupFailed(_ error: String) {
         print("setupFailed: " + error)
     }
@@ -117,14 +137,14 @@ extension AppleASRViewController: SpeechEventListener, PipelineDelegate {
     
     func activate() {
         print("activate")
-        self.stopRecordingButton.isEnabled.toggle()
-        self.startRecordingButton.isEnabled.toggle()
+        self.toggleStartStop()
+        self.pipeline?.activate()
     }
     
     func deactivate() {
         print("deactivate")
-        self.stopRecordingButton.isEnabled.toggle()
-        self.startRecordingButton.isEnabled.toggle()
+        self.toggleStartStop()
+        self.pipeline?.deactivate()
     }
     
     func didError(_ error: Error) {
@@ -137,12 +157,10 @@ extension AppleASRViewController: SpeechEventListener, PipelineDelegate {
     
     func didStart() {
         print("didStart")
-        self.stopRecordingButton.isEnabled.toggle()
-        self.startRecordingButton.isEnabled.toggle()
+        self.toggleStartStop()
     }
     
     func didTrace(_ trace: String) {
         print("didTrace: \(trace)")
     }
 }
-
