+++ SpokestackFrameworkExample/Detect.mlmodel
