+++ SpokestackFrameworkExample/Encode.lite
