+++ SpokestackFrameworkExample/Filter.mlmodel
