+++ SpokestackFrameworkExample/Info.plist
