+++ SpokestackTests/Info.plist
