+++ SpokestackTests/WebRTCVADTest.swift
@@ -1,13 +1,13 @@
 //
 //  WebRTCVADTest.swift
-//  SpokeStackTests
+//  SpokestackTests
 //
 //  Created by Noel Weichbrodt on 9/3/19.
 //  Copyright © 2019 Pylon AI, Inc. All rights reserved.
 //
 
 import XCTest
-import SpokeStack
+import Spokestack
 
 class WebRTCVADTest: XCTestCase {
